\name{readBPM}
\alias{readBPM}
\alias{readBPM}

\title{Low-level function to read BPM files}

\description{
    Reads BPM files (array design information) from Illumina arrays.
}

\usage{
readBPM(file)
}

\arguments{
  \item{file}{character string specifying idat file to be read in}
}

\value{
  A list.
}

\author{Henrik Bengstsson}
\keyword{IO}
